requirements.txt
requests==2.24.0
urllib3==1.25.10
python-twitter
beautifulsoup4
termcolor
numpy
